implemented input port (message listener)
incremental #11
incremental #13

a big refactor on event firing has also been done. now all events must
implemente a `fire` method that just publishes the event. for that, all
events will have a new parameter in the constructor. usages of these
constructors in domain services (in domain-core modules) forces us to
propagate the publisher parameter.

since the domain services in domain-core are used in the domain app
services, the actual publishers will be injected there as beans.